% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_operator.R
\name{f_operator}
\alias{f_operator}
\title{function operator}
\usage{
f_operator(f)
}
\arguments{
\item{f}{a function name}
}
\value{
a new function
}
\description{
function operator
}
\examples{
f_operator(mysqrt)
}
