% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysqrt.R
\name{mysqrt}
\alias{mysqrt}
\title{Self defined sqrt function}
\usage{
mysqrt(x)
}
\arguments{
\item{x}{nonnegative value}
}
\value{
sqrt of x
}
\description{
Self defined sqrt function
}
\examples{
x=2
mysqrt(x)
}
